💥 Fix type of bottle-on-wall example
Migration Guide:

The example previously had the array elements as numbers. The reason the original test cases had that was because they were generated from a `<input type="number" />`. Since this information is not provided in the test input, we should not expect the field to be turned into a number.